Modify vertex reader to use GarDataSource
